added secure routes, burger menu is not available if you are not logged in
