+++ src/SVG/BurgerMenuSVG.jsx
@@ -1,0 +1,20 @@
+import React from 'react';
+
+export const BurgetMenuSVG = () => {
+    return (
+        <svg
+            xmlns='http://www.w3.org/2000/svg'
+            width='50'
+            height='50'
+            fill='currentColor'
+            className='bi bi-list'
+            viewBox='0 0 16 16'
+            style={{ color: 'white' }}
+        >
+            <path
+                fillRule='evenodd'
+                d='M2.5 12a.5.5 0 0 1 .5-.5h10a.5.5 0 0 1 0 1H3a.5.5 0 0 1-.5-.5zm0-4a.5.5 0 0 1 .5-.5h10a.5.5 0 0 1 0 1H3a.5.5 0 0 1-.5-.5zm0-4a.5.5 0 0 1 .5-.5h10a.5.5 0 0 1 0 1H3a.5.5 0 0 1-.5-.5z'
+            />
+        </svg>
+    );
+};
